Reorganize email functions
