+++ reporting-anything/src/report_helpers.js
@@ -2,16 +2,11 @@
  * Copyright OpenSearch Contributors
  * SPDX-License-Identifier: Apache-2.0
  */
-import hbs from "nodemailer-express-handlebars";
-import AWS from "aws-sdk";
-AWS.config.update({region: 'REGION'});
-AWS.config = new AWS.Config();
-const ses = new AWS.SES({region: "us-west-2"});
+import { sendEmail } from './email_helpers.js';
 import puppeteer from 'puppeteer-core';
 import fs from 'fs';
 import { JSDOM } from 'jsdom';
 import { CHROMIUM_PATH, FORMAT, REPORT_TYPE, SELECTOR } from './constants.js';
-import nodemailer from "nodemailer";
 import {exit} from "process";
 const BASIC_AUTH = 'basic';
 const COGNITO_AUTH = 'cognito';
@@ -218,36 +213,6 @@ const getUrl = async (url) =>{
   return urlRef;
 };
 
-const sendEmail = async (fileName, sender, recipient, format) => {
-
-  let mailOptions = getmailOptions(format, sender, recipient, fileName);
-
-  console.log('Creating SES transporter');
-  let transporter = nodemailer.createTransport({
-    SES: ses
-  });
-  
-  transporter.use("compile",hbs({
-    viewEngine:{
-      partialsDir:"./views/",
-      defaultLayout:""
-    },
-    viewPath:"./views/",
-    extName:".hbs"
-  }));
-  
-  // send email
-  await transporter.sendMail(mailOptions, function (err, info) {
-    if (err) {
-      console.log(err);
-      console.log('Error sending email');
-
-    } else {
-      console.log('Email sent successfully');
-    }
-  });
-}
-
 const basicAuthentication = async (page, overridePage, url, username, password) => {
   await page.goto(url, { waitUntil: 'networkidle0' });
   console.log('basic authenticating');
@@ -328,36 +293,3 @@ export const readStreamToFile = async (
     console.log('Downloaded report');
   })
 };
-
-function getmailOptions(format, sender, recipient, fileName, mailOptions = {}) {
-  if(format === FORMAT.PNG) {
-    mailOptions = {
-      from: sender,
-      subject: 'This is an email containing your dashboard report',
-      to: recipient,
-      attachments: [
-        {
-          filename: fileName,
-          path: fileName,
-          cid: 'report'
-        }],
-      //specify the template here
-      template : 'index'
-    };
-  } else {
-    mailOptions = {
-      from: sender,
-      subject: 'This is an email containing your dashboard report',
-      to: recipient,
-      attachments: [
-        {
-          filename: fileName,
-          path: fileName,
-          contentType: 'application/pdf'
-        }],
-      //specify the template here
-      template : 'index'
-    };
-  }
-  return mailOptions;
-}
